requirements.txt
pyserial
spotipy
requests